# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: 云裁 MES / WIP API
  version: 1.0.0
  summary: 服装制造“一扎一码”生产管理接口契约
  description: |
    面向管理后台、工人扫码端、打印服务和外部集成的 REST API。

    核心约束：
    - 所有工厂业务请求必须携带 `X-Factory-Id`，服务端同时执行 RBAC 与数据范围校验。
    - 所有写请求必须携带 `Idempotency-Key`；相同键和相同请求体返回首次结果，相同键但不同请求体返回 `409 IDEMPOTENCY_KEY_REUSED`。
    - 生产事实时间以服务器时间为准，客户端时间仅用于弱网和离线审计。
    - 完工数量必须满足 `inputQty = goodQty + defectQty + missingQty`。
    - 完工、步骤更新、下一步骤解锁、计件快照、事件和审计在同一事务内提交。
    - 已完成生产事实不可覆盖；更正、返工和计件调整均以追加记录实现。
    - 金额以十进制定点字符串传输，禁止使用 IEEE 754 浮点数。
  contact:
    name: MES Product & Engineering
  license:
    name: Proprietary
servers:
  - url: https://mes.example.com
    description: 生产环境示例
  - url: http://127.0.0.1:3000
    description: 本地开发
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: 查询生产订单
      x-permissions: [order:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: orderStatus
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - name: customerId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: styleId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: dueFrom
          in: query
          schema: { type: string, format: date }
        - name: dueTo
          in: query
          schema: { type: string, format: date }
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, -createdAt, dueDate, -dueDate, orderNo, -orderNo]
      responses:
        '200':
          description: 订单列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderPage' }
    post:
      tags: [Orders]
      operationId: createOrder
      summary: 创建生产订单
      x-permissions: [order:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreate' }
      responses:
        '201':
          description: 订单已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/FactoryId'
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      operationId: getOrder
      summary: 获取订单详情与进度
      x-permissions: [order:read]
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Orders]
      operationId: updateOrder
      summary: 更新草稿订单
      description: 已生成扎包后不得修改款式或删除明细，需走变更流程。
      x-permissions: [order:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderPatch' }
      responses:
        '200':
          description: 订单已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/orders/{orderId}/bundle-work-details:
    get:
      tags: [Orders, Bundles]
      operationId: getOrderBundleWorkDetails
      summary: 获取整张订单的扎包执行明细
      description: 仅当订单下全部有效扎包均已完成时返回，用于统一导出员工、工艺、数量和计件金额。
      x-permissions: [bundle:trace]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: 整张订单全部已完成扎包的执行明细
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderBundleWorkDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/cutting-beds:
    get:
      tags: [Cutting]
      operationId: listCuttingBeds
      summary: 查询裁床床次
      x-permissions: [cutting:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - name: orderId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: cuttingStatus
          in: query
          schema: { $ref: '#/components/schemas/CuttingBedStatus' }
        - name: cutDateFrom
          in: query
          schema: { type: string, format: date }
        - name: cutDateTo
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 床次列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CuttingBedPage' }
    post:
      tags: [Cutting]
      operationId: createCuttingBed
      summary: 创建裁床床次
      x-permissions: [cutting:write]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CuttingBedCreate' }
      responses:
        '201':
          description: 床次已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CuttingBed' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/cutting-beds/{cuttingBedId}:
    parameters:
      - $ref: '#/components/parameters/FactoryId'
      - $ref: '#/components/parameters/CuttingBedId'
    get:
      tags: [Cutting]
      operationId: getCuttingBed
      summary: 获取床次及扎包汇总
      x-permissions: [cutting:read]
      responses:
        '200':
          description: 床次详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CuttingBed' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Cutting]
      operationId: updateCuttingBed
      summary: 更新未释放床次
      x-permissions: [cutting:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CuttingBedPatch' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CuttingBed' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/piecework-entries:
    get:
      tags: [Piecework]
      operationId: listPieceworkEntries
      summary: 查询计件事实明细
      x-permissions: [piecework:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - name: workerId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: settlementStatus
          in: query
          schema: { $ref: '#/components/schemas/PieceworkStatus' }
        - name: bundleNo
          in: query
          description: 按扎包号模糊筛选。
          schema: { type: string, maxLength: 80 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 计件明细
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PieceworkEntryPage' }
...
  /api/v1/me/piecework:
    get:
      tags: [Piecework]
      operationId: getMyPiecework
      summary: 查看本人本次、今日、本周和本月计件
      description: 工人只能查看自己的金额；返回预计、已确认和已结算分组。
      x-permissions: [self:piecework:read]
      parameters:
        - $ref: '#/components/parameters/FactoryId'
        - name: period
          in: query
          schema:
            type: string
            enum: [TODAY, WEEK, MONTH, CUSTOM]
            default: TODAY
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 本人计件汇总和明细
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyPieceworkView' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 短期访问令牌；账号停用、设备解绑和会话撤销应立即失效。
  parameters:
    FactoryId:
      name: X-Factory-Id
      in: header
      required: true
      description: 当前业务工厂。服务端必须校验其属于令牌允许的数据范围。
      schema: { $ref: '#/components/schemas/Uuid' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID。相同键和请求体返回首次响应；相同键配不同请求体返回 409。
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源乐观锁版本，格式为带引号的整数 ETag，例如 `"3"`。
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1 }
      description: 上一页返回的不透明游标。
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Search:
      name: q
      in: query
      schema: { type: string, maxLength: 100 }
      description: 编码、名称或业务号模糊查询。
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    CuttingBedId:
      name: cuttingBedId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
  headers:
    XRequestId:
      description: 服务端请求追踪 UUID。
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: 资源不存在；跨工厂资源也统一返回 404，避免信息泄露
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 状态冲突、乐观锁失败、唯一约束冲突或幂等键误用
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: 字段、数量守恒或业务校验失败
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://mes.example.com/problems/quantity-unbalanced
            title: 完工数量不平
            status: 422
            code: QUANTITY_UNBALANCED
            detail: 投入 10，良品、次品和短缺合计 9，还差 1 件。
            requestId: a99a99c1-89dd-4dac-b87d-77d8ea9d74d1
            serverTime: '2026-08-29T11:00:00Z'
            errors:
              - field: missingQty
                code: SUM_MISMATCH
                message: 请补录 1 件。
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Money:
      type: string
      pattern: '^-?[0-9]{1,10}(\.[0-9]{1,4})?$'
      examples: ['3.2000']
    NonNegativeMoney:
      type: string
      pattern: '^[0-9]{1,10}(\.[0-9]{1,4})?$'
      examples: ['0.8000']
    Quantity:
      type: integer
      minimum: 0
    PositiveQuantity:
      type: integer
      minimum: 1
...
    OrderStatus:
      type: string
      enum: [DRAFT, RELEASED, IN_PROGRESS, COMPLETED, CANCELLED]
    CuttingBedStatus:
      type: string
      enum: [DRAFT, CUTTING, CUT, RELEASED, CANCELLED]
    BundleStatus:
      type: string
      enum: [CREATED, IN_PROGRESS, BLOCKED, COMPLETED, CANCELLED, SPLIT, MERGED]
...
    PieceworkStatus:
      type: string
      enum: [PENDING, CONFIRMED, SETTLED, REVERSED]
...
    Problem:
      type: object
      required: [type, title, status, code, requestId, serverTime]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code:
          type: string
          description: 稳定的机器可读错误码。
          examples: [QUANTITY_UNBALANCED]
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        requestId: { type: string, format: uuid }
        serverTime: { type: string, format: date-time }
        retryable: { type: boolean, default: false }
        errors:
          type: array
          items:
            type: object
            required: [code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
              rejectedValue: {}
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
        total:
          type: [integer, 'null']
          minimum: 0
          description: 仅在查询成本可接受时返回。
    AuditFields:
      type: object
      required: [createdAt, updatedAt]
      properties:
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        updatedBy: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        version: { type: integer, minimum: 1 }
...
    OrderItemInput:
      type: object
      required: [lineNo, colorId, sizeId, plannedQty]
      properties:
        lineNo: { type: integer, minimum: 1 }
        colorId: { $ref: '#/components/schemas/Uuid' }
        sizeId: { $ref: '#/components/schemas/Uuid' }
        dyeLotNo: { type: [string, 'null'], maxLength: 60 }
        plannedQty: { $ref: '#/components/schemas/PositiveQuantity' }
        overproductionLimit: { type: integer, minimum: 0, default: 0 }
    OrderItem:
      allOf:
        - $ref: '#/components/schemas/OrderItemInput'
        - type: object
          required: [id, orderId, colorCode, colorName, sizeCode, sizeName, allocatedQty]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            orderId: { $ref: '#/components/schemas/Uuid' }
            colorCode: { type: string }
            colorName: { type: string }
            sizeCode: { type: string }
            sizeName: { type: string }
            allocatedQty: { $ref: '#/components/schemas/Quantity' }
            completedQty: { $ref: '#/components/schemas/Quantity' }
    OrderCreate:
      type: object
      required: [orderNo, styleId, items]
      properties:
        orderNo: { type: string, minLength: 1, maxLength: 60 }
        customerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        styleId: { $ref: '#/components/schemas/Uuid' }
        plannedStartDate: { type: [string, 'null'], format: date }
        dueDate: { type: [string, 'null'], format: date }
        externalRef: { type: [string, 'null'], maxLength: 100 }
        notes: { type: [string, 'null'], maxLength: 2000 }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderItemInput' }
    OrderPatch:
      type: object
      minProperties: 1
      properties:
        customerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        styleId: { $ref: '#/components/schemas/Uuid' }
        plannedStartDate: { type: [string, 'null'], format: date }
        dueDate: { type: [string, 'null'], format: date }
        externalRef: { type: [string, 'null'], maxLength: 100 }
        notes: { type: [string, 'null'], maxLength: 2000 }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderItemInput' }
    Order:
      allOf:
        - type: object
          required: [id, factoryId, orderNo, styleId, styleCode, status, totalPlannedQty, items]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            factoryId: { $ref: '#/components/schemas/Uuid' }
            orderNo: { type: string }
            customerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            customerName: { type: [string, 'null'] }
            styleId: { $ref: '#/components/schemas/Uuid' }
            styleCode: { type: string }
            styleName: { type: string }
            status: { $ref: '#/components/schemas/OrderStatus' }
            plannedStartDate: { type: [string, 'null'], format: date }
            dueDate: { type: [string, 'null'], format: date }
            totalPlannedQty: { $ref: '#/components/schemas/PositiveQuantity' }
            allocatedQty: { $ref: '#/components/schemas/Quantity' }
            completedQty: { $ref: '#/components/schemas/Quantity' }
            progressPercent: { type: string, pattern: '^[0-9]{1,3}(\.[0-9]{1,2})?$' }
            externalRef: { type: [string, 'null'] }
            notes: { type: [string, 'null'] }
            items:
              type: array
              items: { $ref: '#/components/schemas/OrderItem' }
        - $ref: '#/components/schemas/AuditFields'
    OrderPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Order' } }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    CuttingBedCreate:
      type: object
      required: [orderId, bedNo, cutDate]
      properties:
        orderId: { $ref: '#/components/schemas/Uuid' }
        bedNo: { type: string, minLength: 1, maxLength: 40 }
        cutDate: { type: string, format: date }
        plyCount: { type: [integer, 'null'], minimum: 1 }
        dyeLotNo: { type: [string, 'null'], maxLength: 60 }
        supervisorWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        notes: { type: [string, 'null'], maxLength: 1000 }
    CuttingBedPatch:
      type: object
      minProperties: 1
      properties:
        cutDate: { type: string, format: date }
        plyCount: { type: [integer, 'null'], minimum: 1 }
        dyeLotNo: { type: [string, 'null'], maxLength: 60 }
        supervisorWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        status: { $ref: '#/components/schemas/CuttingBedStatus' }
        notes: { type: [string, 'null'], maxLength: 1000 }
    CuttingBed:
      allOf:
        - type: object
          required: [id, factoryId, orderId, bedNo, cutDate, status, bundleCount, totalBundleQty]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            factoryId: { $ref: '#/components/schemas/Uuid' }
            orderId: { $ref: '#/components/schemas/Uuid' }
            orderNo: { type: string }
            bedNo: { type: string }
            cutDate: { type: string, format: date }
            plyCount: { type: [integer, 'null'] }
            dyeLotNo: { type: [string, 'null'] }
            status: { $ref: '#/components/schemas/CuttingBedStatus' }
            supervisorWorkerId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
            bundleCount: { type: integer, minimum: 0 }
            totalBundleQty: { $ref: '#/components/schemas/Quantity' }
            notes: { type: [string, 'null'] }
        - $ref: '#/components/schemas/AuditFields'
    CuttingBedPage:
      type: object
      required: [items, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CuttingBed' } }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    OrderBundleWorkDetailSummary:
      type: object
      required: [id, orderNo, status, styleCode, styleName, bundleCount, completedBundleCount, totalEffectiveQty, totalCompletedQty]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        orderNo: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        styleCode: { type: string }
        styleName: { type: string }
        bundleCount: { type: integer, minimum: 1 }
        completedBundleCount: { type: integer, minimum: 1 }
        totalEffectiveQty: { $ref: '#/components/schemas/Quantity' }
        totalCompletedQty: { $ref: '#/components/schemas/Quantity' }
    OrderBundleWorkDetail:
      type: object
      required: [order, bundles]
      properties:
        order: { $ref: '#/components/schemas/OrderBundleWorkDetailSummary' }
        bundles:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/BundleWorkDetail' }
    BundleWorkDetailSummary:
      type: object
      required: [id, bundleNo, status, bedNo, orderNo, styleCode, styleName, colorCode, colorName, sizeCode, sizeName, dyeLotNo, effectiveQty, completedQty, routeVersionNo]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        bundleNo: { type: string }
        status: { $ref: '#/components/schemas/BundleStatus' }
        bedNo: { type: string }
        orderNo: { type: string }
        styleCode: { type: string }
        styleName: { type: [string, 'null'] }
        colorCode: { type: string }
        colorName: { type: string }
        sizeCode: { type: string }
        sizeName: { type: string }
        dyeLotNo: { type: [string, 'null'] }
        effectiveQty: { $ref: '#/components/schemas/Quantity' }
        completedQty: { $ref: '#/components/schemas/Quantity' }
        routeVersionNo: { type: integer, minimum: 1 }
    BundleWorkDetailRow:
      type: object
      required: [stepNo, processCode, processName, workerNo, workerName, status, startedAt, completedAt, inputQty, goodQty, defectQty, missingQty, unitRate, amount, notes]
      properties:
        stepNo: { type: integer, minimum: 1 }
        processCode: { type: string }
        processName: { type: string }
        workerNo: { type: [string, 'null'] }
        workerName: { type: [string, 'null'] }
        status: { type: string }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        inputQty: { $ref: '#/components/schemas/Quantity' }
        goodQty: { $ref: '#/components/schemas/Quantity' }
        defectQty: { $ref: '#/components/schemas/Quantity' }
        missingQty: { $ref: '#/components/schemas/Quantity' }
        unitRate: { $ref: '#/components/schemas/NonNegativeMoney' }
        amount: { $ref: '#/components/schemas/NonNegativeMoney' }
        notes: { type: [string, 'null'] }
    BundleWorkDetail:
      type: object
      required: [bundle, rows]
      properties:
        bundle: { $ref: '#/components/schemas/BundleWorkDetailSummary' }
        rows:
          type: array
          items: { $ref: '#/components/schemas/BundleWorkDetailRow' }
...
    PieceworkEntry:
      type: object
      required: [id, factoryId, workReportId, workerId, workerNo, workerName, bundleId, bundleNo, processId, processName, quantity, unitRate, amount, status, occurredAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        factoryId: { $ref: '#/components/schemas/Uuid' }
        workReportId: { $ref: '#/components/schemas/Uuid' }
        workerId: { $ref: '#/components/schemas/Uuid' }
        workerNo: { type: string }
        workerName: { type: string }
        bundleId: { $ref: '#/components/schemas/Uuid' }
        bundleNo: { type: string }
        processId: { $ref: '#/components/schemas/Uuid' }
        processName: { type: string }
        quantity: { type: integer }
        unitRate: { $ref: '#/components/schemas/NonNegativeMoney' }
        amount: { $ref: '#/components/schemas/Money' }
        status: { $ref: '#/components/schemas/PieceworkStatus' }
        isRework: { type: boolean }
        adjustmentOfId: { oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }] }
        reason: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }
        confirmedAt: { type: [string, 'null'], format: date-time }
        settledAt: { type: [string, 'null'], format: date-time }
    PieceworkSummary:
      type: object
      required: [totalAmount, totalQuantity, workerCount, bundleCount, entryCount]
      properties:
        totalAmount: { $ref: '#/components/schemas/Money' }
        totalQuantity: { type: integer }
        workerCount: { type: integer, minimum: 0 }
        bundleCount: { type: integer, minimum: 0 }
        entryCount: { type: integer, minimum: 0 }
    PieceworkEntryPage:
      type: object
      required: [items, summary, page]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PieceworkEntry' } }
        summary: { $ref: '#/components/schemas/PieceworkSummary' }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    MyPieceworkView:
      type: object
      required: [workerId, period, estimatedAmount, confirmedAmount, settledAmount, items, page]
      properties:
        workerId: { $ref: '#/components/schemas/Uuid' }
        period:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: date }
            to: { type: string, format: date }
        estimatedAmount: { $ref: '#/components/schemas/Money' }
        confirmedAmount: { $ref: '#/components/schemas/Money' }
        settledAmount: { $ref: '#/components/schemas/Money' }
        goodQty: { $ref: '#/components/schemas/Quantity' }
        defectQty: { $ref: '#/components/schemas/Quantity' }
        items:
          type: array
          items: { $ref: '#/components/schemas/PieceworkEntry' }
        page: { $ref: '#/components/schemas/PageInfo' }
...